Remove rounding and cropping
There's currently no way to tell what is being cropped out
this results in peoples heads being removed, or images
being unclear - such as national flags.

Rounding is not yet working for all image sizes, as it is
applied via a layer mask on a bounding rectangle. Removed
until we can find a way to apply it consistently.